Debug

+++ Unsubscribe.php
@@ -1,5 +1,9 @@
 <?php
 session_start();
+if(!isset($_SESSION['username'])){
+    echo "You are not allow to this page";
+    exit();
+}
 $username=$_SESSION['username'];
 require_once 'function.php';
 queryMysql("DELETE FROM db339.member WHERE member.username='$username'");
